+++ pubtools/pulplib/__init__.py
@@ -12,6 +12,7 @@ from ._impl.model import (
     Unit,
     FileUnit,
     RpmUnit,
+    Dependency,
     ErratumUnit,
     ErratumReference,
     ErratumPackageCollection,