+++ pubtools/pulplib/_impl/fake/rpmlib.py
@@ -25,4 +25,75 @@ except Exception as ex:  # pragma: no cover, pylint: disable=broad-except
     get_header_fields = broken
     get_keys_from_header = broken
 
-__all__ = ["get_rpm_header", "get_header_fields", "get_keys_from_header"]
+__all__ = [
+    "get_rpm_header",
+    "get_header_fields",
+    "get_keys_from_header",
+    "get_rpm_requires",
+    "get_rpm_provides",
+]
+
+
+def get_rpm_requires(header):
+    header_flags_map = {
+        "version": "REQUIREVERSION",
+        "name": "REQUIRENAME",
+        "flags": "REQUIREFLAGS",
+    }
+    return _get_rpm_deps(header, header_flags_map)
+
+
+def get_rpm_provides(header):
+    header_flags_map = {
+        "version": "PROVIDEVERSION",
+        "name": "PROVIDENAME",
+        "flags": "PROVIDEFLAGS",
+    }
+    return _get_rpm_deps(header, header_flags_map)
+
+
+def _get_rpm_deps(header, header_flags_map):
+    header_flags = list(header_flags_map.values())
+    raw_deps = get_header_fields(header, header_flags)
+
+    names = raw_deps[header_flags_map["name"]]
+    versions = raw_deps[header_flags_map["version"]]
+    flags = raw_deps[header_flags_map["flags"]]
+
+    deps = []
+
+    for name, vr, flag in zip(names, versions, flags):
+        version, release = vr.split("-")
+
+        deps_item = {
+            "name": name,
+            "version": version,
+            "release": release,
+            "epoch": None,  # unsure how to get this value from rpm headers
+            "flags": _parse_dep_relation(flag),
+        }
+
+        deps.append(deps_item)
+
+    return deps
+
+
+def _parse_dep_relation(flag):
+    LT = 0x02
+    GT = 0x04
+    EQ = 0x08
+
+    if flag & LT and flag & EQ:
+        flag_str = "LE"
+    elif flag & GT and flag & EQ:
+        flag_str = "GE"
+    elif flag & LT:
+        flag_str = "LT"
+    elif flag & GT:
+        flag_str = "GT"
+    elif flag & EQ:
+        flag_str = "EQ"
+    else:
+        flag_str = ""
+
+    return flag_str