+++ pubtools/pulplib/_impl/model/__init__.py
@@ -19,6 +19,7 @@ from .unit import (
     ErratumPackage,
     ErratumModule,
     RpmUnit,
+    Dependency,
     ModulemdUnit,
     ModulemdDefaultsUnit,
     YumRepoMetadataFileUnit,