+++ pubtools/pulplib/_impl/model/unit/__init__.py
@@ -1,6 +1,6 @@
 from .base import Unit, type_ids_for_class
 from .file import FileUnit
-from .rpm import RpmUnit
+from .rpm import RpmUnit, Dependency
 from .erratum import (
     ErratumUnit,
     ErratumReference,