+++ tests/fake/test_fake_upload_rpm.py
@@ -3,7 +3,7 @@ import os
 
 import pytest
 
-from pubtools.pulplib import FakeController, RpmUnit, YumRepository
+from pubtools.pulplib import FakeController, RpmUnit, YumRepository, Dependency
 
 
 @pytest.mark.parametrize("use_file_object", [False, True])
@@ -57,5 +57,24 @@ def test_can_upload_units(data_path, use_file_object):
             sha256sum="e837a635cc99f967a70f34b268baa52e0f412c1502e08e924ff5b09f1f9573f2",
             content_type_id="rpm",
             repository_memberships=["repo1"],
+            requires=[
+                Dependency(
+                    name="rpmlib(CompressedFileNames)",
+                    version="3.0.4",
+                    release="1",
+                    flags="LE",
+                ),
+                Dependency(
+                    name="rpmlib(PayloadFilesHavePrefix)",
+                    version="4.0",
+                    release="1",
+                    flags="LE",
+                ),
+            ],
+            provides=[
+                Dependency(
+                    name="walrus", version="5.21", release="1", epoch=None, flags="EQ"
+                )
+            ],
         )
     ]