+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- n/a
+- Added `requires` and `provides` attrs to `RpmUnit` class
 
 ## [2.18.0] - 2021-11-29